In the process . . .

+++ include/ft_printf.h
@@ -20,15 +20,15 @@ int		ft_printf(const char *format, ...);
 int		ft_formats(va_list args, const char format);
 int		ft_printchar(int c);
 int		ft_printstr(char *str);
-int		ft_print_ptr(unsigned long ptr);
+int		ft_print_ptr(unsigned long long ptr);
 int		ft_printnbr(int n);
 int		ft_print_unsigned(unsigned int n);
 int		ft_print_hex(unsigned int num, const char format);
 int		ft_printpercent(void);
 
 void	ft_putstr(char *str);
-void	ft_put_ptr(intptr_t num);
-int		ft_ptr_len(intptr_t num);
+void	ft_put_ptr(unsigned long long num);
+int		ft_ptr_len(unsigned long long num);
 char	*ft_itoa(int n);
 int		ft_num_len(int n);
 void	ft_put_hex(unsigned int num, const char format);
