+++ src/ft_printf.c
@@ -29,7 +29,7 @@ int	ft_formats(va_list args, const char format)
 	else if (format == 's')
 		print_length += ft_printstr(va_arg(args, char *));
 	else if (format == 'p')
-		print_length += ft_print_ptr(va_arg(args, unsigned long));
+		print_length += ft_print_ptr(va_arg(args, long long));
 	else if (format == 'd' || format == 'i')
 		print_length += ft_printnbr(va_arg(args, int));
 	else if (format == 'u')
